Refactored MetadataSearch to JsonParser, MetadataItemPermissionManager, and MetadataValidation since MetadataSearch was handling multiple tasks. MetadataSearch should now only handle finding/updating metadata items;
Updating and adding tests for the refactored classes;
